+++ Sources/Player/Common/Types/BroadcastPlaybackInfo.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-struct BroadcastPlaybackInfo: Decodable {
-	let id: String
-	let audioQuality: AudioQuality
-	let manifestType: String
-	let manifest: String
-}