+++ Sources/Player/PlaybackEngine/Internal/AVQueuePlayer/AVPlayerItemMonitor.swift
@@ -7,7 +7,6 @@ final class AVPlayerItemMonitor {
 	private let stallMonitor: StallMonitor
 	private let completelyDownloadedMonitor: CompletelyDownloadedMonitor
 	private let readyToPlayMonitor: ReadyToPlayMonitor
-	private let inStreamMetadataMonitor: DJSessionTransitionMonitor
 	private let playerItemDidPlayToEndTimeMonitor: ItemPlayedToEndMonitor
 
 	init(
@@ -16,8 +15,7 @@ final class AVPlayerItemMonitor {
 		onStall: @escaping (AVPlayerItem) -> Void,
 		onCompletelyDownloaded: @escaping (AVPlayerItem) -> Void,
 		onReadyToPlayToPlay: @escaping (AVPlayerItem) -> Void,
-		onItemPlayedToEnd: @escaping (AVPlayerItem) -> Void,
-		onDjSessionTransition: @escaping (AVPlayerItem, DJSessionTransition) -> Void
+		onItemPlayedToEnd: @escaping (AVPlayerItem) -> Void
 	) {
 		failureMonitor = FailureMonitor(playerItem, onFailure)
 		failedToPlayToEndMonitor = FailedToPlayToEndMonitor(playerItem, onFailure)
@@ -25,11 +23,6 @@ final class AVPlayerItemMonitor {
 		completelyDownloadedMonitor = CompletelyDownloadedMonitor(playerItem, onCompletelyDownloaded)
 		readyToPlayMonitor = ReadyToPlayMonitor(playerItem, onReadyToPlayToPlay)
 		playerItemDidPlayToEndTimeMonitor = ItemPlayedToEndMonitor(playerItem: playerItem, onPlayedToEnd: onItemPlayedToEnd)
-		inStreamMetadataMonitor = DJSessionTransitionMonitor(
-			with: playerItem,
-			and: DispatchQueue.global(qos: .default),
-			onTransition: onDjSessionTransition
-		)
 	}
 
 	func isCompletelyDownloaded() -> Bool {