+++ Sources/Player/PlaybackEngine/Internal/Broadcasting/CreateBroadcastResponse.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-struct CreateBroadcastResponse: Decodable {
-	let broadcastId: String
-	let curationUrl: URL
-	let sharingUrl: URL
-	let reactions: [String]?
-}