+++ Sources/Player/PlaybackEngine/Internal/PlayerItem.swift
@@ -13,7 +13,6 @@ protocol PlayerItemMonitor: AnyObject {
 	func downloaded(playerItem: PlayerItem)
 	func completed(playerItem: PlayerItem)
 	func failed(playerItem: PlayerItem, with error: Error)
-	func djSessionTransition(playerItem: PlayerItem, transition: DJSessionTransition)
 	func playbackMetadataLoaded(playerItem: PlayerItem)
 }
 
@@ -301,14 +300,6 @@ extension PlayerItem: PlayerMonitoringDelegate {
 
 		metrics?.recordEnd(endReason: .ERROR, assetPosition: asset.getAssetPosition(), error: error)
 		playerItemMonitor?.failed(playerItem: self, with: error)
-	}
-
-	func djSessionTransition(asset: Asset?, transition: DJSessionTransition) {
-		guard asset === self.asset else {
-			return
-		}
-
-		playerItemMonitor?.djSessionTransition(playerItem: self, transition: transition)
 	}
 
 	func playbackMetadataLoaded(asset: Asset?) {