+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Restore `ProductType.UC` for local file playback while keeping remote UC handling removed (Player)
 
+### Removed
+- Removed broadcast/DJ functionality from the Player module (Player)
+
 ## [0.7.7] - 2025-10-14
 
 ### Changed